refactor(retrieval-api): extract ranking helpers to api/ranking.py
Behavior-preserving lift of the three pure scoring/ranking helpers
(_rrf_merge, _apply_link_expand_boost, _compute_confidence_band) out of
retrieve.py into a cohesive sibling. All three are side-effect-free over the
in-flight chunk lists and already have dedicated unit tests. retrieve re-imports
them so the orchestrator call sites and the
`from retrieval_api.api.retrieve import ...` test imports are unchanged.

retrieval-api suite: 530 passed, 1 skipped. ruff clean.
